Create Tic-Tac-Toe Game
I heard Bill Gates say that the first program he made was a Tic-Tac-Toe game.  I didn't really believe him, because most people's first program is something like "add 1+1" or "prompt the user for their name," so I set out to see if I could build a game of my own!

It should be noted that the computer opponent plays randomly in this game.  I could have built an AI into the game, but that was beyond the scope of this project.  Additionally, a perfect AI will never lose in Tic-Tac-Toe, so it seemed reasonable to make the computer dumb.