Create Blackjack
This is an incomplete version of a Blackjack game I've been working on.  As it stands the player draws but the dealer does not, hence there is no competitive aspect.  I plan to include that later.

Also, I'm not sure I got all the rules of Blackjack right.  I've never played it before!